Moved GsapPosition to gsap package. Added Svelte easing functions to gsap.
