+++ src/gsap/gsap.js
@@ -1,3 +1,5 @@
+import * as easingFuncs from 'svelte/easing';
+
 let gsap = void 0;
 
 const modulePath = foundry.utils.getRoute('/scripts/greensock/esm/all.js');
@@ -6,6 +8,13 @@ try
 {
    const module = await import(modulePath);
    gsap = module.gsap;
+
+   // Load Svelte easing functions by prepending them w/ `svelte-`; `linear` becomes `svelte-linear`, etc.
+   for (const prop of Object.keys(easingFuncs))
+   {
+      const name = `svelte-${prop}`;
+      gsap.registerEase(name, easingFuncs[prop]);
+   }
 }
 catch (error)
 {
