+++ src/gsap/index.js
@@ -1,4 +1,6 @@
 export * from './gsap.js';
+
 export * from './animate.js';
+export * from './GsapPosition.js';
